Get rid of temporary inspector panel components.
